Moved dotfiles from $HOME to $HOME/.dotfiles
I like to keep as little files as neccessary in my $HOME dir.
When I can specify the config file for a program. Eg: GNU/Screen
it is better to alias the command to include the config file rather
than keep the config file in the $HOME dir.

Moved, shell_aliases, and screenrc.